+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kopflos-cms/core
 
+## 0.3.3
+
+### Patch Changes
+
+- 7aa216f: `KopflosEnvironment` is also added as `this` context of handler factory
+
+## 0.3.2
+
+### Patch Changes
+
+- dbc9de5: Added `handlers.js` module with `getCoreRepresentation` handler
+
 ## 0.3.1
 
 ### Patch Changes